Extract urlFragmentFromState into urlBuilder.
This prepares me to make the two branches (Collect vs Collection) different.
